moved services/ to integrations/, fixed tests
