+++ pkg/model/auth.go
@@ -28,9 +28,12 @@ func NewPharosRepoAuth(authDsn string, tlsCheck bool) (PharosRepoAuth, error) {
 
 // return true if auth is not empty and matchies imageRef repo
 func (rx PharosRepoAuth) HasAuth(imageRef string) bool {
-	if rx.Authority != "" && (rx.Username != "" || rx.Token != "") {
-		if strings.HasPrefix(imageRef, rx.Authority) {
-			return true
+
+	if rx.Authority != "" {
+		if rx.Username != "" || rx.Token != "" {
+			if strings.HasPrefix(imageRef, rx.Authority) {
+				return true
+			}
 		}
 	}
 
@@ -40,10 +43,10 @@ func (rx PharosRepoAuth) HasAuth(imageRef string) bool {
 // return DSN without password
 func (rx PharosRepoAuth) ToMaskedDsn(mask string) string {
 	if rx.Password != "" {
-		return fmt.Sprintf("registry://%s:%s@%s/?type=password", rx.Username, mask, rx.Authority)
+		return fmt.Sprintf("registry://%s:%s@%s", rx.Username, mask, rx.Authority)
 	}
 	if rx.Token != "" {
-		return fmt.Sprintf("registry://%s:%s@%s/?type=token", rx.Username, mask, rx.Authority)
+		return fmt.Sprintf("registry://%s:%s@%s", rx.Token, "", rx.Authority)
 	}
 	return ""
 }
@@ -59,28 +62,29 @@ func (rx *PharosRepoAuth) FromDsn(input string) error {
 	rx.Password = ""
 	rx.Token = ""
 
+	// no input is avlid to streamline code flow in calls
 	if input == "" {
 		return nil
 	}
 	dsn := dsnparser.Parse(input)
 	if dsn == nil {
-		return fmt.Errorf("invalid registry %v", input)
+		return fmt.Errorf("invalid registry dsn: %s", input)
 	}
 	// build authority
 	rx.Authority = dsn.GetHost()
 	if dsn.GetPort() != "" {
 		rx.Authority = dsn.GetHost() + ":" + dsn.GetPort()
 	}
-	what := dsn.GetParam("type")
-	if what == "password" {
+	fmt.Printf("auth u=%s, p=%s, h=%s, p=%s\n", dsn.GetUser(), dsn.GetPassword(), dsn.GetHost(), dsn.GetPort())
+
+	if dsn.GetUser() != "" && dsn.GetPassword() != "" {
 		rx.Username = dsn.GetUser()
 		rx.Password = dsn.GetPassword()
 		return nil
-	} else if what == "token" {
-		rx.Token = dsn.GetPassword()
-		return nil
-	} else {
-		return fmt.Errorf("invalid DSN type '%s' (e.g. 'registry://usr:pwd@docker.io/?type=password')", what)
 	}
-
+	if dsn.GetUser() != "" {
+		rx.Token = dsn.GetUser()
+		return nil
+	}
+	return fmt.Errorf("invalid auth dsn: %s", input)
 }
