+++ scanner/cmd/scan.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	"github.com/dustin/go-humanize"
-	"github.com/metraction/pharos/internal/services/cache"
+	"github.com/metraction/pharos/internal/integrations/cache"
 	"github.com/metraction/pharos/pkg/grype"
 	"github.com/metraction/pharos/pkg/model"
 	"github.com/metraction/pharos/pkg/trivy"
